requirements.txt
httplib2==0.9
python-instagram==1.3.0
simplejson==3.6.5
six==1.9.0
